Castle capture in place. Move on top of the enemy castle to win the game. Briefly tested to make sure both player and CPU can win via castle capture.
At this point the console version of the game is basically done. I'd prefer it be tested more extensively before this branch is pulled to master.

The only things that should probably be added are:
Prompt user for AI difficulty (1-3, anything above 3 takes forever)
Prompt user for map size

These are both hard coded into the game, and honestly I haven't tried changing the map size. To be honest, for demo purposes I don't care. I don't even care if we do something larger, I doubt they'll hold us to it, and 20x20 is a nice size. AI as well. But if the group decides we should stick to what we had promised, then these features should just be added in along with the UI anyway.